Controller structure improved.
AppContainer -> BaseController
AppEnv separated from AppContainer(BaseController)

The purpose of BaseController is that we could easily reuse functions
that are generic and used widely across the project. There is one
function that isn't (GetPostModel) that isn't generic at the moment
i'll have to update go version to go1.18 across the board so we could
make the function GetPostModel reusable.

We should figure out how we could create the Database connection
in MakeHandler and close the function in MakeHanlder as well.

I'm not sure how we could achieve this.